Add files via upload
This is a short video of the robot gripper moving. The test code runs all the servos, but my computer doesn’t supply enough power. Because of that, the gripper stalls under load, and some of the servos twitch power brownout). I’ll be getting some 18650 batteries soon to fix that.